changed int parameter to size_t in ast_iface_new()
which probably had little to no effect whatsoever, except to keep it more consistent, and avoid compiler warnings

+++ src/ast.c
@@ -67,7 +67,7 @@ ast_domain_new(char *domain, char *url)
 }
 
 struct ast_iface *
-ast_iface_new(char *name, int ndomains)
+ast_iface_new(char *name, size_t ndomains)
 {
 	assert(name);
 	assert(ndomains > 0);
